Backend update

+++ backend/MedicalPlant/Model/Plant.cs
@@ -6,14 +6,14 @@ public class Plant
     public string LatinName { get; set; }
     public string CommonName { get; set; }
     public string ArabicName { get; set; }
-    public int OriginId { get; set; }
-    public int VegetableReignId { get; set; } // Régne
-    public int PlantDivisionId { get; set; }
-    public int PlantClassId { get; set; }
-    public int PlantFamilyId { get; set; }
-    public int PlantGenreId { get; set; }
-    public int PlantSpeciesId { get; set; }
-    public int UsedPartId { get; set; }
+    public long OriginId { get; set; }
+    public long VegetableReignId { get; set; } // Régne
+    public long PlantDivisionId { get; set; }
+    public long PlantClassId { get; set; }
+    public long PlantFamilyId { get; set; }
+    public long PlantGenreId { get; set; }
+    public long PlantSpeciesId { get; set; }
+    public long UsedPartId { get; set; }
     
     // Upon removal of plants, we keep them in the database
     // With Removed boolean set to true
